+++ src/utils/helpers.js
@@ -2,8 +2,13 @@ import {
   Quiz as QuizIcon,
   Leaderboard as LeaderboardIcon,
   Person as PersonIcon,
+  School as SchoolIcon,
+  Security as SecurityIcon,
+  Payment as PaymentIcon,
+  Settings as SettingsIcon,
   Terminal
 } from '@mui/icons-material';
+
 /**
  * Performance Optimization Utilities
  * Helper functions for optimizing React and Material-UI applications
@@ -261,4 +266,58 @@ export const AdminDashboardRoutes = [
     iconBg: 'warning.main',
     navigateTo: '/server-logs',
   },
+];
+
+export const categories = [
+    {
+      icon: <SchoolIcon sx={{ fontSize: 40 }} />,
+      title: 'Getting Started',
+      description: 'Learn the basics of using QuizApp',
+      color: '#667eea',
+    },
+    {
+      icon: <SecurityIcon sx={{ fontSize: 40 }} />,
+      title: 'Account & Privacy',
+      description: 'Manage your account and privacy settings',
+      color: '#10b981',
+    },
+    {
+      icon: <PaymentIcon sx={{ fontSize: 40 }} />,
+      title: 'Billing & Plans',
+      description: 'Information about pricing and payments',
+      color: '#f59e0b',
+    },
+    {
+      icon: <SettingsIcon sx={{ fontSize: 40 }} />,
+      title: 'Features & Settings',
+      description: 'Explore features and customize settings',
+      color: '#ec4899',
+    },
+];
+
+export  const faqs = [
+    {
+      question: 'How do I create an account?',
+      answer: 'Click on the "Get Started" button on the homepage and fill in your details. You\'ll receive a confirmation email to verify your account.',
+    },
+    {
+      question: 'Is TriviaVerse free to use?',
+      answer: 'Yes! TriviaVerse offers a free plan with access to thousands of quizzes. We also have premium plans with additional features for power users.',
+    },
+    {
+      question: 'How do I create my own quiz?',
+      answer: 'After logging in, please drop an email to update your role to "Creator" then navigate to the Dashboard and click on "Create Quiz". Follow the step-by-step wizard to add questions, set time limits, and publish your quiz. More details will be available in the Tutorial footer menu!',
+    },
+    {
+      question: 'Can I track my progress?',
+      answer: 'Absolutely! Your Dashboard provides detailed analytics including scores, time spent, improvement trends, and areas for improvement.',
+    },
+    {
+      question: 'How does the leaderboard work?',
+      answer: 'The leaderboard ranks users based on their quiz performance, including accuracy, speed, and consistency. Points are calculated using our proprietary algorithm.',
+    },
+    {
+      question: 'What if I encounter a technical issue?',
+      answer: 'Please contact our support team through the Contact page or email us at triviaverse.contact@gmail.com. We typically respond within 24 hours.',
+    },
 ];
